Adds many worlds content display
Implements the display of world, category, and entity data.

Introduces WorldController with methods for creating, retrieving, and editing worlds, categories, and entities.
Adds routes for accessing and manipulating this data.
Adds Segment entity with display order and active status.
Updates WorldEntity entity to include entityImgUrl.

The front end is updated to display the many-worlds content with proper navigation, and an admin user can edit content.